requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
Deprecated==1.2.18
distro==1.9.0
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
openai==1.97.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyGithub==2.6.1
Pygments==2.19.2
PyJWT==2.10.1
PyNaCl==1.5.0
python-dotenv==1.1.1
requests==2.32.4
rich==14.1.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
wrapt==1.17.2
